requirements.txt
aspose.slides
aspose.pydrawing